pyproject.toml
[tool.poetry]
name = "test_project_miss_x"
version = "0.1.0"
description = ""
authors = ["Yaroslav Galiulin <yaroslav66@list.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
tinydb = "^4.8.0"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
